Update requirements_webui.txt
Fix ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
google-ai-generativelanguage 0.6.6 requires protobuf!=3.20.0,!=3.20.1,!=4.21.0,!=4.21.1,!=4.21.2,!=4.21.3,!=4.21.4,!=4.21.5,<5.0.0dev,>=3.19.5, but you have protobuf 5.28.0 which is incompatible.
gradio 4.37.1 requires urllib3~=2.0, but you have urllib3 1.26.20 which is incompatible.
openxlab 0.1.1 requires packaging~=24.0, but you have packaging 23.2 which is incompatible.
openxlab 0.1.1 requires requests~=2.28.2, but you have requests 2.32.3 which is incompatible.
openxlab 0.1.1 requires tqdm~=4.65.0, but you have tqdm 4.66.5 which is incompatible.